Moving elasticsearch directory into the translation directory and renaming to hubmap_translation. Updating translation_helper_function class to avoid circular referencing
