+++ src/translator/hubmap_translation/addl_index_transformations/portal/cache/.gitignore
@@ -1,3 +1,3 @@
-*
+README.md
 !README.md
 !.gitignore
