+++ src/translator/hubmap_translation/addl_index_transformations/portal/reset_entity_type.py
