+++ src/translator/hubmap_translation/addl_index_transformations/portal/tests/fixtures/input-doc.json
