+++ src/translator/hubmap_translation/neo4j-to-es-attributes.json
